requirements.txt
azure-storage-blob==12.19.0
pandas==2.1.3
watchdog==3.0.0